+++ guicey-server-pages/src/test/groovy/ru/vyarus/guicey/gsp/views/RendererDetectionTest.groovy
@@ -15,10 +15,6 @@ import spock.lang.Specification
  */
 class RendererDetectionTest extends Specification {
 
-    void cleanup() {
-        ServerPagesBundle.resetGlobalConfig()
-    }
-
     def "Check renderer requirement check"() {
 
         when: "starting app"